Add code to make the openvpn-gui.exe usable under win8
Under win8, the openvpn-gui is not run as an admin and may not
appear in the systray, albeit it is actually running. To remedy this,
one can set the openvpn gui to be run as admin and also set it to be
run in e.g. win7 compatibility mode. With this commit, there is code
for the installer to detect the location of the openvpn-gui and then
write a value in the registry that will tell the system to actually
run openvpn-gui with admin privileges and in compatibility mode.